% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_warc.r
\name{as_warc}
\alias{as_warc}
\title{Convert an \code{httr::respone} object to WARC response objects}
\usage{
as_warc(x)
}
\arguments{
\item{x}{\code{httr} \code{response} object}
}
\description{
Convert an \code{httr::respone} object to WARC response objects
}

