% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_copy_dataframe_labels.R
\name{copy_dataframe_labels}
\alias{copy_dataframe_labels}
\title{Copy labels between dataframes}
\usage{
copy_dataframe_labels(to, from)
}
\arguments{
\item{to}{(Dataframe) The dataframe whose columns will be labelled.}

\item{from}{(Dataframe) The dataframe whose labels will be used.}
}
\value{
A version of \code{to} with variables labelled. Since \code{Hmisc} is used
to do the labelling, it also adds the \code{labelled} class to the vector.
}
\description{
This uses \code{\link[Hmisc:label]{Hmisc::label<-()}} to apply labels to the dataframe. Dataframes that were
labelled by \code{\link[Hmisc:label]{Hmisc::label()}}, \code{\link[haven:labelled]{haven::labelled()}}, or \code{\link[labelled:var_label]{labelled::var_label()}} are
accepted.
}
\details{
Value labels are also copied over.

Column names must exactly match between \code{to} and \code{from}. Columns in \code{to} that
are not in \code{from} will not be modified.
}
\examples{
x <- copy_dataframe_labels(iris, iris_labelled_hmisc)
str(x)

## data.frame':	150 obs. of  5 variables:
##     $ Sepal.Length: 'labelled' num  5.1 4.9 4.7 4.6 5 5.4 4.6 5 4.4 4.9 ...
## ..- attr(*, "label")= chr "Length of sepals."
## $ Sepal.Width : 'labelled' num  3.5 3 3.2 3.1 3.6 3.9 3.4 3.4 2.9 3.1 ...
## ..- attr(*, "label")= chr "Width of sepals."
## $ Petal.Length: 'labelled' num  1.4 1.4 1.3 1.5 1.4 1.7 1.4 1.5 1.4 1.5 ...
## ..- attr(*, "label")= chr "Length of petals."
## $ Petal.Width : 'labelled' num  0.2 0.2 0.2 0.2 0.2 0.4 0.3 0.2 0.2 0.1 ...
## ..- attr(*, "label")= chr "Width of petals."
## $ Species     : Factor w/ 3 levels "setosa","versicolor",..: 1 1 1 1 1  ...
## ..- attr(*, "label")= chr "Species of iris: setosa, versicolor, virginica"

}
