% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_funcs.R
\name{copy_definition}
\alias{copy_definition}
\title{Copy factor levels, ordering, and labels from one vector to another}
\usage{
copy_definition(to, from, ...)
}
\arguments{
\item{to}{(Vector) The vector to be factored and labelled.}

\item{from}{(Vector) The vector whose factor levels and labels will be used.}

\item{...}{(...) Additional arguments that will be passed to \code{\link[=factor]{factor()}}.}
}
\value{
A factorised version of \code{to} with levels, level ordering, and labels (if any).
}
\description{
This function supports vectors that were labelled with either \code{\link[Hmisc:label]{Hmisc::label()}} or
\code{\link[haven:labelled]{haven::labelled()}}.
}
\details{
Labels will only be applied to \code{to} if they existed in \code{from}.
}
\examples{
# ---- Labelled with Hmisc ----

fac_old_ord <-
    ordered(c("Always", "Sometimes", "Always", "Rarely",
              "Never", "Sometimes", "Rarely", "Never"),
            levels = c("Always", "Sometimes", "Rarely", "Never"))

Hmisc::label(fac_old_ord) <- "This label is for my ordered and labelled factor."

fac_new_ord <- c("Rarely", "Always", "Sometimes", "Never",
                 "Never", "Always", "Rarely")


# This copies the levels, ordering, and label from `fac_old_ord` to `fac_new_ord`.
copy_definition(fac_new_ord, fac_old_ord)

#> An example of an ordered and labelled factor.
#> [1] Rarely    Always    Sometimes Never     Never     Always    Rarely
#> [8] Sometimes
#> Levels: Always < Sometimes < Rarely < Never



# ---- Labelled with haven ----

haven_old <- haven::labelled(c("M", "M", "F"),
                             c(Male = "M", Female = "F"),
                             "This label is for my haven-labelled factor.")

haven_new <- c("M", "F", "F", "F", "M")


copy_definition(haven_new, haven_old)

#> <labelled<character>[5]>: This label is for my haven-labelled factor.
#> [1] M F F F M
#>
#> Labels:
#>     value  label
#> M   Male
#> F Female



# ---- Unlabelled ----

fac_old_unord <-
    factor(c("Curtain", "Cat", "Keyboard",
             "Keyboard", "Cat", "Curtain"))

fac_new_unord <- c("Curtain", "Keyboard", "Curtain",
                   "Cat", "Cat", "Keyboard")


# Since `fac_old_unord` was unlabelled, the output is also unlabelled.
copy_definition(fac_new_unord, fac_old_unord)

#> [1] Curtain  Keyboard Curtain  Cat      Cat      Keyboard
#> Levels: Cat Curtain Keyboard

}
