% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_funcs.R
\name{diff_cols}
\alias{diff_cols}
\title{Columns that don't appear in two dataframes}
\usage{
diff_cols(df1, df2, side = "both")
}
\arguments{
\item{df1}{(Dataframe) A dataframe whose column names to compare.}

\item{df2}{(Dataframe) A dataframe whose column names to compare.}

\item{side}{(Character) \code{"both"} or \code{"b"} (default) finds columns that are missing from
both dataframes. \code{"left"} or \code{"l"} finds cols in \code{df1} that are not in \code{df2}.
\code{"right"} or \code{"r"} finds cols in \code{df2} that are not in \code{df1}.}
}
\value{
A Character vector with the names of missing columns.
}
\description{
Columns that don't appear in two dataframes
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
iris1 <- iris[, 1:3]
colnames(iris1)
## [1] "Sepal.Length" "Sepal.Width"  "Petal.Length"

iris2 <- iris[, 2:5]
colnames(iris2)
## [1]                "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"

diff_cols(iris1, iris2)
#> [1] "Sepal.Length" "Petal.Width"  "Species"

diff_cols(iris1, iris2, side = "l")
#> [1] "Sepal.Length"

diff_cols(iris1, iris2, side = "r")
#> [1] "Petal.Width"  "Species"

}
