% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{expand}
\alias{expand}
\title{Expand a compressed buffer}
\usage{
expand(buffer, size = length(buffer) * 10)
}
\arguments{
\item{buffer}{buffer}

\item{size}{size}
}
\description{
Expand a compressed buffer
}

