% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_sequence}
\alias{find_sequence}
\title{Find the first occurrence (if any) of a sequence of raw bytes
(\code{pattern}) in \code{buffer}.}
\usage{
find_sequence(buffer, pattern)
}
\arguments{
\item{buffer}{vector to search in}

\item{pattern}{sequence of bytes to look for}
}
\value{
index in \code{buffer} or \code{-1} if not found
}
\description{
Find the first occurrence (if any) of a sequence of raw bytes
(\code{pattern}) in \code{buffer}.
}

