% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gz_flush}
\alias{gz_flush}
\title{Flush currenzt gzip stream}
\usage{
gz_flush(gzfile)
}
\arguments{
\item{gzfile}{file handle}
}
\description{
This will flush all zlib output buffers for the current file
and terminate the gzip stream. The next \code{gz_write()} will
start a new gzip stream.
}

