% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gz_fseek}
\alias{gz_fseek}
\title{Sets the starting position for the next \code{gz_read()} or \code{gz_write()}}
\usage{
gz_fseek(gzfile, offset, from)
}
\arguments{
\item{gzfile}{file handle}

\item{offset}{represents a number of bytes in the compressed data stream}

\item{from}{either "\code{start}", "\code{end}" or "\code{current}"}
}
\value{
\code{TRUE} if successful
}
\description{
Sets the starting position for the next \code{gz_read()} or \code{gz_write()}
}

