% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gz_gets_raw}
\alias{gz_gets_raw}
\title{Read a line from a gz file}
\usage{
gz_gets_raw(gzfile)
}
\arguments{
\item{gzfile}{file handle}
}
\description{
Read a line from a gz file
}
\note{
line buffer max is 8,192 characters. The intent of this function is to use it
  on well-known formats.
}

