% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gz_offset}
\alias{gz_offset}
\title{Return the current raw compressed offset in the file}
\usage{
gz_offset(gzfile)
}
\arguments{
\item{gzfile}{file handle}
}
\value{
offset position (integer)
}
\description{
Return the current raw compressed offset in the file
}

