% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gz_open}
\alias{gz_open}
\title{Open a gzip file for reading or writing}
\usage{
gz_open(path, mode)
}
\arguments{
\item{path}{path to file}

\item{mode}{"\code{write}" or "\code{read}"}
}
\value{
handle to the file
}
\description{
Open a gzip file for reading or writing
}

