% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gz_read_raw}
\alias{gz_read_raw}
\title{Read from a gz file into a raw vector}
\usage{
gz_read_raw(gzfile, len)
}
\arguments{
\item{gzfile}{file handle}

\item{len}{number of of characters}
}
\description{
Read from a gz file into a raw vector
}

