% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gz_seek}
\alias{gz_seek}
\title{Sets the starting position for the next \code{gz_read()} or \code{gz_write()}}
\usage{
gz_seek(gzfile, offset, from)
}
\arguments{
\item{gzfile}{file handle}

\item{offset}{represents a number of bytes in the uncompressed data stream}

\item{from}{either "\code{start}" or "\code{current}"}
}
\value{
the resulting offset location as measured in bytes from the beginning of the
  uncompressed stream, or –1 in case of error, in particular if the file is opened
  for writing and the new starting position would be before the current position.
}
\description{
Sets the starting position for the next \code{gz_read()} or \code{gz_write()}
}

