% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gz_tell}
\alias{gz_tell}
\title{Return the current raw uncompressedf offset in the file}
\usage{
gz_tell(gzfile)
}
\arguments{
\item{gzfile}{file handle}
}
\value{
offset position (integer)
}
\description{
Return the current raw uncompressedf offset in the file
}

