% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gz_write_char}
\alias{gz_write_char}
\title{Write an atomic character vector to a file}
\usage{
gz_write_char(gzfile, buffer)
}
\arguments{
\item{gzfile}{file handle}

\item{buffer}{atomic character vector}
}
\description{
Write an atomic character vector to a file
}

