% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gz_write_raw}
\alias{gz_write_raw}
\title{Write a raw vector to a gz file}
\usage{
gz_write_raw(gzfile, buffer)
}
\arguments{
\item{gzfile}{file handle}

\item{buffer}{raw vector to write}
}
\description{
Write a raw vector to a gz file
}

