% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{iris_labelled_haven}
\alias{iris_labelled_haven}
\title{The classic Iris dataset, labelled with 'haven'}
\format{
\subsection{\code{iris_labelled_haven}}{

A data frame with 150 rows and 5 columns:
\describe{
\item{Sepal.Length}{Length of sepals.}
\item{Sepal.Width}{Width of sepals.}
\item{Petal.Length}{Length of petals.}
\item{Petal.Width}{Width of petals.}
\item{Species}{Species of iris: setosa, versicolor, virginica}
...
}
}
}
\source{
R \code{datasets} package
}
\usage{
iris_labelled_haven
}
\description{
This was labelled with \code{\link[haven:labelled]{haven::labelled()}}.
}
\keyword{datasets}
