% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdx.r
\name{read_cdx}
\alias{read_cdx}
\title{Read a WARC CDX index file}
\usage{
read_cdx(path, warc_path = NULL)
}
\arguments{
\item{path}{path to CDX file}

\item{warc_path}{path to the WARC files referenced in \code{path}. Defaults to
to the location of the CDX file}
}
\description{
CDX files are used to index the content of WARC files.
}
\details{
The returned object is a \code{tbl_df} but is also classed
\code{cdx}.
}
\examples{
\dontrun{
cdx <- read_cdx(system.file("extdata", "20160901.cdx", package="warc"))
}
}
\references{
\url{https://iipc.github.io/warc-specifications/specifications/cdx-format/cdx-2015/}
}

