% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warc.R
\name{read_warc}
\alias{read_warc}
\title{Read a WARC file}
\usage{
read_warc(path, entry_filter = "", line_filter = "")
}
\arguments{
\item{entry_filter}{a character string that, when not found inside the entry, will
cause the entry to be filtered out.}

\item{line_filter}{a character string that, when not found inside a line, will
cause that line to be filtered out.}

\item{The}{path to the WARC file.}
}
\description{
Reads a WARC file from the given path and optionally applies filters
per entry or line.
}
