% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_report_dims.R
\name{report_dims}
\alias{report_dims}
\title{Report the difference between the number of rows in two dataframes}
\usage{
report_dims(x, y)
}
\arguments{
\item{x}{(Dataframe) A dataframe to compare.}

\item{y}{(Dataframe) A dataframe to compare.}
}
\value{
Prints to the console.
}
\description{
Report the difference between the number of rows in two dataframes
}
\details{
This is a symmetric comparison; the order of \code{x} and \code{y} doesn't matter.
}
\examples{
report_dims(head(iris, 10), iris)
}
