% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_report_setdiff.R
\name{report_setdiff}
\alias{report_setdiff}
\title{Compare sets of values between two vectors or columns}
\usage{
report_setdiff(x, y)
}
\arguments{
\item{x}{(Vector) Vector to compare.}

\item{y}{(Vector) Vector to compare.}
}
\value{
Prints to the console. Invisibly returns a named list (\code{x_diff_from_y}
and \code{y_diff_from_x}) of differing values.
}
\description{
This is a wrapper around \code{\link[=setdiff]{setdiff()}} that a) performs the set difference in
both directions, b) reports the differences in a way that's easy to understand
("x has 4 values not in y", and vice versa), and c) returns the vectors of
differences.
}
\examples{
set.seed(12345)
df1 <- data.frame(patient    = letters[1:10],
                  hospital   = sample(c("WH", "NH", "CH"), 10, replace = TRUE))

df2 <- data.frame(patient_id = letters[5:26],
                  hospital   = sample(c("WH", "NH", "CH"), 22, replace = TRUE))

report_setdiff(df1$hospital, df2$hospital)
report_setdiff(df1$patient,  df2$patient_id)
}
