% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_funcs.R
\name{same_cols}
\alias{same_cols}
\title{Columns that appear in two dataframes}
\usage{
same_cols(df1, df2)
}
\arguments{
\item{df1}{(Dataframe) A dataframe whose column names to compare.}

\item{df2}{(Dataframe) A dataframe whose column names to compare.}
}
\value{
A Character vector with the names of the columns that appear
in both \code{df1} and \code{df2}.
}
\description{
Columns that appear in two dataframes
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
iris1 <- iris[, 1:3]
colnames(iris1)
## [1] "Sepal.Length" "Sepal.Width"  "Petal.Length"

iris2 <- iris[, 2:5]
colnames(iris2)
## [1]                "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"

same_cols(iris1, iris2)
#> [1] "Sepal.Width"  "Petal.Length"

}
