% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_funcs.R
\name{shrink_redcap_cols}
\alias{shrink_redcap_cols}
\title{Remove duplicate columns that are generated by REDCap's data import code}
\usage{
shrink_redcap_cols(df)
}
\arguments{
\item{df}{(Dataframe) A dataframe imported using REDCap's auto-generated R code.}
}
\value{
A dataframe.
}
\description{
REDCap generates R code to import data and label it, which is great. But when it does
labels and levels factor columns, it makes a copy of the original column (for example,
\code{original} copied into \code{original.factor}) that doesn't even have the labels from the
original, which is bad.
}
\details{
This function:
\enumerate{
\item Copies the labels from \code{original} into \code{original.factor},
\item removes \code{original},
\item renames \code{original.factor} to \code{original}.
}
}
