% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonumber.R
\name{tbl_number}
\alias{tbl_number}
\alias{fig_number}
\alias{map_number}
\title{Auto-numbering of tables, figures, and maps}
\usage{
tbl_number(i = NULL)

fig_number(i = NULL)

map_number(i = NULL)
}
\arguments{
\item{i}{(Character) Keywords that control which number is printed.
\itemize{
\item An empty argument or anything \strong{not} in this keyword list will add 1 to the number.
\item To return the current item number, use \code{"curr"} or \code{"this"}.
\item To return the last item number, use \code{"prev"} or \code{"last"}.
\item To return the next item number, use \code{"next"}.
}}
}
\value{
A Character vector of length one.
}
\description{
These functions insert caption numbers ("Table 1", "Figure 1", etc.) that automatically
increment every time the function is called.
}
\details{
Each function stores its own count, so use the correct function for the
correct item type.
}
\section{Functions}{
\itemize{
\item \code{tbl_number()}: Auto-numbering of Tables.

\item \code{fig_number()}: Auto-numbering of Figures.

\item \code{map_number()}: Auto-numbering of Maps.

}}
\examples{
my_table_caption <- paste(tbl_number(), "Summary statistics for iris petal lengths.")
my_table_caption
#> [1] "Table 1. Summary statistics for iris petal lengths."

another_table <- paste(tbl_number(), "Another table.")
another_table
#> [1] "Table 2. Another table."

# What is the current table number?
tbl_number("curr")
#> [1] "Table 2."

# What was the table number before this one?
tbl_number("prev")
#> [1] "Table 1."

# What is the number of the next table?
tbl_number("next")
#> [1] "Table 3."

# Each autonumber function tracks its own count.
fig_number()
#> [1] "Figure 1."

map_number()
map_number()
#> [1] "Map 2."

}
