% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_warc_entry.r
\name{warc_headers}
\alias{warc_headers}
\title{Extract WARC headers from a WARC response object}
\usage{
warc_headers(x)
}
\arguments{
\item{x}{WARC response object}
}
\value{
list of WARC headers
}
\description{
Extract WARC headers from a WARC response object
}

