% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_deprecated.R
\name{with_label}
\alias{with_label}
\title{Deprecated: One-step evaluation and labelling}
\usage{
with_label(expr, label)
}
\arguments{
\item{expr}{(Expression) An expression to evaluate.}

\item{label}{(Character) The label to apply to the output.}
}
\value{
The evaluated result of \code{expr}, labelled with \code{label}.
}
\description{
This function is being deprecated because applying labels inside a
dplyr pipeline works inconsistently. Mostly it works fine, but
sometimes the labels don't get applied at all, and I can't discern
a reason why. Use \code{\link[labelled:var_label]{labelled::set_variable_labels()}} instead.
}
