% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_warc.r
\name{write_warc_record}
\alias{write_warc_record}
\title{Write a WARC record to a file}
\usage{
write_warc_record(warc_rec, warc_path, append = TRUE)
}
\arguments{
\item{warc_rec}{a WARC Record (as created by \code{as_warc})}

\item{warc_path}{path to WARC file}

\item{append}{if \code{TRUE} then the data will be appended to \code{warc_path}}
}
\description{
Write a WARC record to a file
}

