% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_warc.r
\name{create_warc}
\alias{create_warc}
\title{Use wget to create a WARC archive for a URL list}
\usage{
create_warc(url_list, warc_path, user_agent = "r-warc", max_redirects = 5,
  tries = 2, waitretry = 1, timeout = 5, warc_header = "r-warc",
  warc_cdx = TRUE, warc_file = "r-warc", no_warc_digests = TRUE,
  no_warc_keep_log = TRUE, warc_max_size = "1G", warc_tempdir = tempdir(),
  .opts = NULL)
}
\arguments{
\item{url_list}{character vector of URLs or a \code{connection} to a file
with one URL per line}

\item{warc_path}{path where to store WARC archive output}

\item{user_agent, max_redirects, tries, waitretry, timeout, warc_header, warc_cdx}{options for \code{wget}}

\item{warc_file, nowarc_digests, no_warc_keep_log, warc_max_size, warc_tempdir}{options for \code{wget}}

\item{.opts}{a \code{list} of other options for \code{wget}. Use long-argument names replacing
\code{-} with \code{_} and using \code{TRUE} or \code{FALSE} for boolean arguments.}
}
\description{
\code{wget} must be available, on the system \code{PATH} and be compiled with
WARC support.
}
\note{
\code{Mozilla/5.0 (Windows; U; Windows NT 6.1; en-US) AppleWebKit/533.20.25 (KHTML, like Gecko) Version/5.0.4 Safari/533.20.27}
}

