% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{expand}
\alias{expand}
\title{Expand a compressed raw buffer}
\usage{
expand(buffer, size = length(buffer) * 10)
}
\arguments{
\item{buffer}{gz compressed buffer to expand}

\item{size}{estimated size of the inflated buffer}
}
\description{
Expand a compressed raw buffer
}
\note{
Guess \code{size} well since this will fail if \code{size}
    is not large enough to hold the uncompressed output.
}

