% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gz_close}
\alias{gz_close}
\title{Close the gz file}
\usage{
gz_close(gzfile, flush)
}
\arguments{
\item{gzfile}{file handle}

\item{flush}{flush zlib buffers for the current file and terminate the gzip stream?}
}
\description{
Close the gz file
}

