% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdx.r
\name{read_cdx}
\alias{read_cdx}
\title{Read a WARC CDX index file}
\usage{
read_cdx(path, warc_path = dirname(path))
}
\arguments{
\item{path}{path to CDX file}

\item{warc_path}{path to the WARC files referenced in \code{path}. Defaults to
to the location of the CDX file}
}
\description{
Read a WARC CDX index file
}
\examples{
\dontrun{
cdx <- read_cdx(system.file("extdata", "20160901.cdx", package="warc"))
}
}
\references{
\url{https://iipc.github.io/warc-specifications/specifications/cdx-format/cdx-2015/}
}

