% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_warc_entry.r
\name{read_warc_entry}
\alias{read_warc_entry}
\title{Read a WARC entry from a WARC file}
\usage{
read_warc_entry(path, start, size, compressed = grepl(".gz$", path),
  compressed_buffer_multiplier = 10)
}
\arguments{
\item{path}{path to WARC file}

\item{start}{starting offset of WARC record}
}
\description{
Given the path to a WARC file (compressed or uncompressed) and the start
positions of the WARC record, this function will produce an R object from the
WARC record.
}
\details{
WARC \code{warinfo} objects are returned classed both \code{warc} and
\code{info}.

WARC \code{response} objects are returned classed both \code{warc} and
\code{httr::response} and the standard \code{httr} content functions
will work with the object.

WARC \code{request} objects are returned classed both \code{warc} and
\code{httr::request}.
}
\note{
\code{warcinfo}, \code{request} and \code{response} objects are currently
  supported.
}
\examples{
\dontrun{
cdx <- read_cdx(system.file("extdata", "20160901.cdx", package="warc"))
i <- 1
path <- file.path(cdx$warc_path[i], cdx$file_name[i])
start <- cdx$compressed_arc_file_offset[i]

(read_warc_entry(path, start))
}
}

