% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_warc_entry.r
\name{read_warc_entry}
\alias{read_warc_entry}
\title{Read a WARC entry from a WARC file}
\usage{
read_warc_entry(path, start, size, compressed = grepl("gz$", path))
}
\arguments{
\item{path}{path to WARC file}

\item{start}{starting offset of WARC record}

\item{size}{total size of WARC record}

\item{compressed}{is WARC file compressed? (guessed by file name suffix)}
}
\description{
Read a WARC entry from a WARC file
}
\examples{
\dontrun{
cdx <- read_cdx(system.file("extdata", "20160901.cdx", package="warc"))
p <- file.path(cdx$warc_path[i], cdx$file_name[i])
st <- cdx$compressed_arc_file_offset[i]
sz <- cdx$calc_compressed_warc_rec_size[i]

cat(read_warc_entry(p, st, sz))
}
}

