% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_report_missingness.R
\name{report_missingness}
\alias{report_missingness}
\title{Report missingness of columns in a dataframe}
\usage{
report_missingness(df, pattern, ...)
}
\arguments{
\item{df}{(Dataframe) A dataframe.}

\item{pattern}{(Character) A regular expression that will be used to select columns of \code{df}, passed to \code{\link[=agrep]{agrep()}}.}

\item{...}{(Dots) Other arguments that will be passed to \code{\link[=agrep]{agrep()}}.}
}
\value{
Prints to the console.
}
\description{
Report missingness of columns in a dataframe
}
\examples{
report_missingness(iris, "^Petal")
report_missingness(iris, "Length$")
}
