% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_autonumber.R
\name{tbl_number}
\alias{tbl_number}
\alias{fig_number}
\alias{map_number}
\alias{appx_number}
\title{Auto-numbering of tables, figures, and maps}
\usage{
tbl_number(i = NULL)

fig_number(i = NULL)

map_number(i = NULL)

appx_number(i = NULL)
}
\arguments{
\item{i}{(\code{NULL} or Numeric)
\itemize{
\item To increment the internal counter, pass \code{NULL} or skip \code{i} (default).
\item Otherwise, to 'look around' at the item numbers \strong{without} incrementing the counter,
pass a number. For example:
\itemize{
\item \code{i = 0} returns the current item number,
\item \code{i = -1} returns the previous one, and
\item \code{i = +3} returns the number 3 items ahead.
}
}}
}
\value{
A Character vector of length one.
}
\description{
These functions insert caption numbers ("Table 1", "Figure 1", etc.) that automatically
increment every time the function is called. Backwards- and forwards-referencing are
also supported.
}
\details{
Since the auto-number increments every time the function is called, this means that
if you keep running the same table repeatedly, then its table number will keep ticking
up. \strong{That's perfectly okay!} When you restart R and source the code (or if you're
knitting a RMarkdown document that uses this auto-numbering), then the numbers will be
correct in the output.

Each function stores its own count, so use the correct function for the
correct item type.
}
\section{Functions}{
\itemize{
\item \code{tbl_number()}: Auto-numbering of Tables.

\item \code{fig_number()}: Auto-numbering of Figures.

\item \code{map_number()}: Auto-numbering of Maps.

\item \code{appx_number()}: Auto-numbering of Appendices.

}}
\examples{
# Assembling a table caption using base R's paste():
paste0(tbl_number(), ". My first table.")
#> [1] "Table 1. My first table."

# Assembling a table caption using base R's sprintf():
sprintf("\%s. My second table", tbl_number())
#> [1] "Table 2. My second table."

# What is the current table number?
tbl_number(0)
#> [1] "Table 2"

# What was the table number before this one?
tbl_number(-1)
#> [1] "Table 1"

# What is the number of the next table?
tbl_number(+1)
#> [1] "Table 3"

# What is the number of the table 10 items from now?
tbl_number(+10)
#> [1] "Table 12"

# Each autonumber function tracks its own count.
fig_number()
#> [1] "Figure 1"

map_number()
#> [1] "Map 1"

map_number()
#> [1] "Map 2"

}
