% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_deprecated.R
\name{warc_dirs}
\alias{warc_dirs}
\title{Deprecated: Generate shortcuts to folders in WARC's standardised folder structure}
\usage{
warc_dirs(rds_folder, quiet = FALSE)
}
\arguments{
\item{rds_folder}{(Character) Path to working directory.}

\item{quiet}{(Character) If \code{FALSE} (default), report shortcuts.}
}
\value{
A named list of Character vectors, which are paths to folders.
}
\description{
This function is removed and should not be used. Use relative paths to
navigate the WARC folder structure instead.

For example, \code{"../../../03_Data/raw.csv"} navigates 3 directory levels
upwards/backwards (\verb{../} 3 times), then enters the folder \verb{03_Data} to access
the file \code{raw.csv}.
}
